Keep finding interactions in their commands
